tests/wpt/web-platform-tests/webmessaging/with-ports/026.html

<!doctype html>
<title>object with length as transferable</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function() {
   assert_throws(new TypeError(), function() {
      postMessage('', '*', {length:1});
   });
});
</script>
